+++ GPTifier/src/api.cpp
@@ -1,4 +1,4 @@
-#include "query.hpp"
+#include "api.hpp"
 #include "params.hpp"
 #include "utils.hpp"
